aumento el tamanio del precio de un prod o accesorio

+++ database/migrations/2020_09_13_193126_create_productos_table.php
@@ -18,7 +18,7 @@ class CreateProductosTable extends Migration
             $table->id();
             $table->foreignId('proveedor_id');
             $table->string('nombre');
-            $table->float('precio')->nullable();
+            $table->bigInteger('precio')->nullable();
             $table->integer('nro_articulo')->nullable();
             $table->timestamps();
             $table->softDeletes();
